Refactor: Naming & Delete Old Files
