--- Assets/Scripts/Runtime/CH3/TRPG/DialogueTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 51b5f3684ba7f054fac52c625f3207a9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
