+++ Assets/Scripts/Runtime/CH3/TRPG/TrpgDialogue.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0873ae33e6d5bf9479d4748f18a8c81b
+guid: 51b5f3684ba7f054fac52c625f3207a9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
