--- Assets/Scripts/Runtime/CH3/TRPG/TrpgLine.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4d408feba991118448e00329bac6e62b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
